Finished initial version of view user info.
Currently simply displays a list of attributes and values.
I think we need to reconsider what we want this to show.
For instance, I don't think we should show UID.
Also still needs to map database constants to appropriate values. I didn't
know how things were converted in the past so I ignored this for now.
I also think my solution wasn't as clean as it should be. Perhaps there is a
better way to pass the information to the template.
I also fixed the static paths to use url_for so that they don't break and
added ordereddict support since we are using python-2.6.