Added government page (unstyled)
New government page to list all current offices. It displays three sets of
offices: excomm, uccs, and other house offices. Additionally, clicking on a row
links to the person's profile (which prompts for log-in).

Currently the page is unstyled, which is ugly. @connorcoley make it pretty? :D
Additionally, the page has a bit of duplicate code. I didn't know how we would
end up organizing the government page, so I decided this was OK for now.

Part of Issue #5.