Bug fix with authentication
We only checked if a user was an office at one point which had
sufficient permissions for viewing a page. However, we should check that
they are CURRENTLY in a position that allows permissions to view a page.
This is a simple fix that uses `office_members_current` instead of
`office_members`.