Fix update mail script
The script to update email lists was broken after the addition of
blueprints/modules in PR #71. This is because the import was incorrect.
As a fix, I've moved the update script out of the
`RuddockHouse/RuddockHouse/` folder. This is also nicer because the
update script isn't run on the website, so it makes sense to have it
outside that directory.