Adds support for moving between buckets.
Creates a new view on the main portal- a dropdown to select two
distinct buckets to compare.

Creates a comparison page which holds two scrollable divs for the two
buckets. Prefrosh can be moved from the left bucket to the right bucket
using the 'move' button. Scroll position is maintained after a move.