Adds support for moving between buckets. (#128)
* Adds support for moving between buckets.

Creates a new view on the main portal- a dropdown to select two
distinct buckets to compare.

Creates a comparison page which holds two scrollable divs for the two
buckets. Prefrosh can be moved from the left bucket to the right bucket
using the 'move' button. Scroll position is maintained after a move.

* Adds back button to bucket moving page.

Adds 0.5 and 1.5 buckets.

* Changes move page to use a get request instead of a post request.

Minor nit cleanup.

* removed print statements