Use user ID for viewing profiles everywhere
The view_profile function rendered a view of a member's information. It
would fetch this information using the user's username. Since we should
be able to view, edit, and remove members regardless of whether they
have accounts on the website, I decided to make a view_member function
instead that does the same thing, except it uses the user's user_id,
which everyone has, regardless of whether they've made an account on the
website or not. (Maybe a better name for this would have been
member_id...)

This commit removes all references to view_profile in favor of
view_member. Doing this involves making some minor changes to queries
in helper functions here and there, as well as adding the user's user_id
to the flask session when they log in (so that the URL to view their
profile in the header is properly generated for every page). Finally, it
changes the view_profile test to test view_member instead.