Convert website to Python 3
This commit consists of a few independent changes. They will be
stated in chronological order.

First, my local tests indicated that some of the SQL was malformed
according to new versions of MariaDB. I changed some and commented
out others. This is orthogonal to the rest of the commit, and can
be taken out if deemed unnecessary.

Second, requirements.txt was changed from using MySQL-Python to
using mysqlclient. mysqlclient is a fork of MySQL-Python that is
Python 3 compliant and doesn't require as many dependencies.

Third, I ran 2to3 on the whole repository, revealing only a few
motifs that required changing. Most involved using "http.client"
instead of httplib2 (or something), using "map" when list
comprehensions could be used, print statements, etc. Simple things
that changed in Py3.

Finally, the tests were not passing after this, so I had to change
exactly two things:

1) Some parts of the codebase were using hashlib (and related
modules). This module requires its inputs to be of type 'bytes',
rather than 'str', so the code stopped working. Since the password
input code (from the Flask side of things) still used 'str', I
wrote a hack to get around this. I capture any 'str' inputs in
the hash functions and convert them immediately to 'bytes'.
Going forward, it would be good to propagate this change all the
way down to the Flask code, but for now, it passes the tests...
and probably works.

2) The test that checks the constitution upstream vs. the local
constitution was failing for the same reason as above; http.content
now is 'bytes' instead of 'str'. I converted this. It's not hacky
because it's now just the correct way to use these modules.

The changes I would like to see, going forward, are:
 - Fixing the above hack to be less hacky
 - Using typing, especially for this str/bytes business
 - Adding more tests
 - Increasing the amount of SQL test data