Revert type-signature of hash function
My previous commits changed the signature of the hash_password
function to accept both str and bytes, converting any errant
str's to bytes's. This commit reverts to the previous functionality
where the function only accepts str. Of course, this entails a lot
of ".encode()" calls, since the hash functions in the libraries
we use require the type to be bytes.

This way is better for typing, and likely cleaner, but slightly
more repetitive. More local variables would assist the repetition,
but that doesn't matter very much at the moment. Ultimately our
functions WOULD eventually take 'bytes', at which point they have
to be rewritten anyway.