Rework the pytest fixtures
There's now 3 fixtures:
  - app: creates the app in test mode; exists for the whole test session.
      likely never imported, but useful for caching.
  - client: mimics a connection to the server. use this for checking web
      requests and responses.
  - transaction: using this automatically wraps your db actions in a
      transaction. use this for tests that do writes to the test db.
      read-only tests don't need this.

By putting them in conftest.py they get auto-discovered. If we don't, then we
have to explicitly import the parent fixtures like "app" which is ridiculous.
I'll admit this loses readability, it's no longer obvious where the fixtures
come from. But on the other hand, the errors that you get by not importing
"app" are genuinely bewildering.

Anyways, I included an example of a transaction-using test, and hopefully it's
clear how to write safe mutation tests :)