Rework the pytest fixtures
There's now 2 fixtures:
  - app: creates the app in test mode
    - useful for testing app configuration and flasky business
  - client: mimics a connection to the server
    - useful for checking web requests and responses.

By putting them in conftest.py they get auto-discovered. If we don't, then we
have to explicitly import the parent fixtures like "app" which is ridiculous.
I'll admit this loses readability, it's no longer obvious where the fixtures
come from. But on the other hand, the errors that you get by not importing
"app" are genuinely bewildering.