Use SqlAlchemy ORM for the budget module
Most of helpers.py was about writing SQL queries in docstrings, and it
turns out that that can be all but eliminated by using the ORM. Wasn't
able to get mypy type checking to work out yet but I've heard it's
possible.

IMO this is much cleaner to read, and the minor performance overhead
doesn't really matter for us, so the tradeoff is worth it.

Also, if we convert all the modules to do this, we can clean up a lot
of the database/*.sql scripts (maybe remove them altogether).